jsconfig.json and friends
Try adding a jsconfig and placeholder.ts to enable intellisense and
fine-tune linting. placeholder.ts is there to stop a linter error based on not
having a TS entrypoint (jsconfig is a TS thing). The .gitignore for foundry
is ignoring an attemp to symlink in the foundry system so intellisense can find it.